# openapi.yaml
openapi: 3.1.0
info:
  title: Bookstore API
  version: 1.0.0
  description: Basic demo e-commerce API
...
paths:
...
  /users/carts:
    description: Returns all carts belonging to the loggedIn user
    get:
      summary: Gets all cart with all its items belonging to the loggedIn user
      description: >-
        Returns a cart with all its items, alongside their title, price, author
        and quantity
      responses:
        "200":
          description: Successful retrieval of a cart
          content:
            application/json:
              examples:
                Successful retrieval of a cart:
                  value: |-
                    [
                        {
                            "title": "Israel-Palestine",
                            "author": "Alain Gresh",
                            "price": "£8.00",
                            "quantity": 3
                        },
                        {
                            "title": "The Return of the King",
                            "author": "J.R.R Tolkien",
                            "price": "£6.99",
                            "quantity": 2
                        }
                    ]
        "400":
          description: Message containing the error stack
        "404":
          description: USer id doesn't match any user in the database
          content:
            text/plain; charset=utf-8:
              examples:
                User not found (invalid user ID):
                  value: User not found
      tags:
        - carts
    post:
      summary: Creates a cart for the loggedIn user
      description: Creates a cart for the loggedIn user
      responses:
        "201":
          description: Successful creation of a cart
          content:
            text/plain; charset=utf-8:
              examples:
                Successful creation of a cart:
                  value: Cart created
        "400":
          description: Message containing the error stack
      tags:
        - carts
  /users/carts/{cartId}:
    description: >-
      Removes a cart, (note that the cart must be empty, in other words, there
      must be no cart items associated to the cart).
    delete:
      summary: Delete a cart, if empty, belonging to the loggedIn user
      parameters:
        - in: path
          name: cartId
          schema:
            type: integer
          required: true
          description: Numeric ID of the cart to delete
      description: >-
        Deletes a cart but also checks if said cart is empty, if not a warning
        will be sent
      responses:
        "204":
          description: Cart successfully deleted
          content:
            text/plain; charset=utf-8:
              examples:
                Successful deletion:
                  value: Cart deleted!
        "400":
          description: Message containing the error stack
        "404":
          description: The cart has cart items associated to it
          content:
            text/plain; charset=utf-8:
              examples:
                Cart is not empty:
                  value: The cart needs to be empty first!
      tags:
        - carts
  /users/carts/{cartId}/checkout:
    description: >-
      Creates an order from the loggedIn user and a cart id, deletes the
      associated cart and its cart items and finally adds the order total to the
      order
    post:
      summary: Checkout a cart for the loggedIn user
      parameters:
        - in: path
          name: cartId
          schema:
            type: integer
          required: true
          description: Numeric ID of the cart to post
      description: >-
        This endpoint creates an order, fills it with each of the cart items
        associated to the user Id and cart ID then deletes said items as well as
        the cart. Finally, it calculates the order total and adds it to the
        order
      responses:
        "200":
          description: Successful checkout
          content:
            text/plain; charset=utf-8:
              examples:
                Successful checkout with 2 items in the cart:
                  value: |
                    Card order added
                    Card order added
                    Cart order cleared!
                    Cart order cleared!
                    Cart deleted
                    Order made
        "400":
          description: Message containing error stack
      tags:
        - carts
  /users/carts/{cartId}/cart_items/{bookId}:
    description: "Gets an item from a cart by the user id, cart id and the book id "
    get:
      summary: Get a cart item by the loggedIn user alongside the cart and book IDs
      parameters:
        - in: path
          name: bookId
          schema:
            type: integer
          required: true
          description: Numeric ID of the book to get
        - in: path
          name: cartId
          schema:
            type: integer
          required: true
          description: Numeric ID of the cart to get
      description: >-
        Gives a cart item which matches, if possible, the given user, cart and
        book IDs with the book's title, author, price and quantity within the
        cart
      responses:
        "200":
          description: Succcesful retrivel of a cart item
          content:
            application/json:
              examples:
                Successful retrieval:
                  value: |-
                    [
                        {
                            "title": "The Return of the King",
                            "author": "J.R.R Tolkien",
                            "price": "£6.99",
                            "quantity": 2
                        }
                    ]
        "400":
          description: Message containing the error stack
        "404":
          description: At least one of the IDs has not matched with a data entry
          content:
            text/plain; charset=utf-8:
              examples:
                At least one of the IDs doesn't match a corresponding entry in the database:
                  value: Cart item not found
      tags:
        - cart_items
    put:
      summary: Update the quantity of a given cart item for a user
      parameters:
        - in: path
          name: bookId
          schema:
            type: integer
          required: true
          description: Numeric ID of the book to put
        - in: path
          name: cartId
          schema:
            type: integer
          required: true
          description: Numeric ID of the  cart to put
      description: Updates the quantity of a cart item whose cart matches the given ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                quantity:
                  type: integer
                  example: 3
                  minimum: 1
      responses:
        "200":
          description: Successful update of the quantity of a card order
          content:
            text/plain; charset=utf-8:
              examples:
                Successful update of a cart item:
                  value: Cart order updated!
        "400":
          description: Message containing the error stack
        "404":
          description: No book matching given ID within the cart's items
          content:
            text/plain; charset=utf-8:
              examples:
                Book not found within a cart's items:
                  value: Book not found
      tags:
        - cart_items
    delete:
      summary: Remove a cart item belonging to the loggedIn user
      parameters:
        - in: path
          name: bookId
          schema:
            type: integer
          required: true
          description: Numeric ID of the book to delete
        - in: path
          name: cartId
          schema:
            type: integer
          required: true
          description: Numeric ID of the cart to delete
      description: Delete a cart item from a given cart
      responses:
        "204":
          description: Successful deletion of a cart item
          content:
            text/plain; charset=utf-8:
              examples:
                Successful deletion of a cart_item:
                  value: "''"
        "400":
          description: Message containing the error stack
      tags:
        - cart_items
  /users/carts/{cartId}/cart_items:
    description: Create a cart item in a given cart for the loggedIn user
    post:
      summary: Creates a cart item within a given cart for the loggedIn user
      parameters:
        - in: path
          name: cartId
          schema:
            type: integer
          required: true
          description: Numeric ID of the cart to post
      description: Creates a cart item in a cart matching the relevant IDs
      responses:
        "201":
          description: Successful creation of a cart item
          content:
            text/plain; charset=utf-8:
              examples:
                Successful creation of a cart item:
                  value: Added to cart!
        "400":
          description: Message containing the error stack
      tags:
        - cart_items
  /orders:
    summary: One of the orders endpoints
    get:
      summary: Retrieves all of a user's orders
      description: Returns all the orders for a given user
      responses:
        "200":
          description: Successful retrieval of orders
          content:
            application/json:
              examples:
                Successful retrieval of orders:
                  value: |-
                    [
                        {
                            "id": 79,
                            "username": "Guest",
                            "email": "guest@gmail.com",
                            "order_total": "£99.96",
                            "created_at": "2022-12-06T14:38:52.058Z"
                        }
                    ]
        "400":
          description: Message containing the error stack
      tags:
        - orders
  /orders/{orderId}:
    summary: Order ID specific endpoints
    get:
      summary: Select a specific order for a user
      parameters:
        - in: path
          name: orderId
          schema:
            type: integer
          required: true
          description: Numeric ID of the order to get
      description: Return an order which matches the given ID
      responses:
        "200":
          description: Successful retrieval of an order
          content:
            application/json:
              examples:
                Successful retrieval of an order (id 79):
                  value: |-
                    [
                        {
                            "id": 79,
                            "user_id": 1,
                            "created_at": "2022-12-06T14:38:52.058Z",
                            "order_total": "£99.96"
                        }
                    ]
        "400":
          description: Message containing the error stack
      tags:
        - orders
...